Admin profile functionalities, edit user, delete user, forms validators

+++ Client/src/app/admin-profile/admin-profile.component.ts
@@ -4,6 +4,9 @@ import { Router, ActivatedRoute } from '@angular/router';
 import { Presentation } from '../models/presentation.model';
 import { PresentationService } from '../services/presentation.service';
 import { UserService } from '../services/user.service';
+import { FormGroup, FormBuilder, Validators } from '@angular/forms';
+import { MustMatch } from '../shared/sign-up.validator';
+import { first } from 'rxjs/operators';
 
 @Component({
   selector: 'app-admin-profile',
@@ -14,17 +17,30 @@ export class AdminProfileComponent implements OnInit {
   users: User[];
   presentations: Presentation[];
   arrEditUserbtn = new Array() as Array<boolean>;
+  submitted = false;
+  loading = false;
+  editUserForm: FormGroup;
+
   constructor(
     private router: Router,
     private route: ActivatedRoute,
     private userService: UserService,
-    private presentationService: PresentationService
+    private presentationService: PresentationService,
+    private formBuilder: FormBuilder
   ) { }
   ngOnInit() {
     this.route.data.subscribe((data: { presentations: Presentation[], users: User[] }) => {
       this.presentations = data.presentations;
       this.users = data.users;
     });
+    this.editUserForm = this.formBuilder.group({
+      email: ['', [Validators.required, Validators.email]],
+      userGroup: ['', Validators.required],
+      password: ['', [Validators.required, Validators.minLength(6)]],
+      confirm_password: ['', Validators.required]
+    }, {
+        validator: MustMatch('password', 'confirm_password')
+      });
   }
   editUser(i: number) {
     console.log(this.users[i].id);
@@ -33,19 +49,68 @@ export class AdminProfileComponent implements OnInit {
     } else {
       this.arrEditUserbtn[i] = true;
     }
+    this.editUserForm = this.formBuilder.group({
+      email: this.users[i].email,
+      userGroup: this.users[i].type,
+      password: this.users[i].password,
+      confirm_password: this.users[i].password
+    }, {
+        validator: MustMatch('password', 'confirm_password')
+      });
+  }
+  onSubmit(i: number) {
+    this.submitted = true;
+
+    // stop here if form is invalid
+    if (this.editUserForm.invalid) {
+      return;
+    }
+
+    this.loading = true;
+    this.userService.updateUser(<User>{
+      id: this.users[i].id,
+      email: this.editUserForm.get('email').value,
+      type: this.editUserForm.get('userGroup').value,
+      password: this.editUserForm.get('password').value
+    }).pipe(first())
+      .subscribe(
+        data => {
+          console.log('Succes');
+          alert('Success');
+          this.submitted = false;
+          this.editUserForm.reset();
+          this.arrEditUserbtn[i] = false;
+          window.location.reload();
+          // this.router.navigate(['/sign-in']);
+        },
+        error => {
+          alert(error);
+          console.log(error);
+          this.loading = false;
+        }
+      );
 
   }
+  onCancel(i: number) {
+    this.arrEditUserbtn[i] = false;
+    this.editUserForm.reset();
+  }
   deleteUser(i: number) {
-    // this.presentationService.deletePresentation(this.presentations[i].id);
+    if (confirm('Are you sure that you want to delete ' + this.users[i].email + ' ?')) {
+      this.userService.deleteUser(this.users[i].id);
+      alert('Deleted');
+    }
   }
   openPresentationPage(i: number) {
-    // console.log('Trying to open presentation ' + this.presentations[i].id);
     this.router.navigate([`/presentation-page/${this.presentations[i].id}`]);
   }
   editPresentationPage(i: number) {
     this.router.navigate([`/edit-presentation/${this.presentations[i].id}`]);
   }
   deletePresentationPage(i: number) {
-    this.presentationService.deletePresentation(this.presentations[i].id);
+    if (confirm('Are you sure that you want to delete ' + this.presentations[i].title + ' presentation ?')) {
+      this.presentationService.deletePresentation(this.presentations[i].id);
+      alert('Deleted');
+    }
   }
 }
